move typeclass instances to typeclass companion object
