+++ core/shared/src/main/scala/monocle/function/Curry.scala
@@ -11,12 +11,31 @@ abstract class Curry[F, G] extends Serializable {
   def curry: Iso[F, G]
 }
 
-object Curry extends CurryFunctions
-
 trait CurryFunctions {
-
   def curry[F, G](implicit ev: Curry[F, G]): Iso[F, G] = ev.curry
-
   def uncurry[F, G](implicit ev: Curry[F, G]): Iso[G, F] = curry.reverse
+}
 
+object Curry extends CurryFunctions with CurryInstances {
+  implicit def curry5[A, B, C, D, E, F] = new Curry[(A, B, C, D, E) => F, A => B => C => D => E => F] {
+    def curry = Iso((_: (A, B, C, D, E) => F).curried)(f => Function.uncurried(f))
+  }
+}
+
+trait CurryInstances extends CurryInstances1 {
+  implicit def curry4[A, B, C, D, E] = new Curry[(A, B, C, D) => E, A => B => C => D => E] {
+    def curry = Iso((_: (A, B, C, D) => E).curried)(f => Function.uncurried(f))
+  }
+}
+
+trait CurryInstances1 extends CurryInstances2 {
+  implicit def curry3[A, B, C, D] = new Curry[(A, B, C) => D, A => B => C => D] {
+    def curry = Iso((_: (A, B, C) => D).curried)(f => Function.uncurried(f))
+  }
+}
+
+trait CurryInstances2 {
+  implicit def curry2[A, B, C] = new Curry[(A, B) => C, A => B => C] {
+    def curry = Iso((_: (A, B) => C).curried)(f => Function.uncurried(f))
+  }
 }
