+++ core/shared/src/main/scala/monocle/function/Field5.scala
@@ -15,14 +15,26 @@ abstract class Field5[S, A] extends Serializable {
   def fifth: Lens[S, A]
 }
 
-object Field5 extends Field5Functions {
-  /** lift an instance of [[Field5]] using an [[Iso]] */
-  def fromIso[S, A, B](iso: Iso[S, A])(implicit ev: Field5[A, B]): Field5[S, B] = new Field5[S, B] {
-    override def fifth: Lens[S, B] =
-      iso composeLens ev.fifth
-  }
-}
-
 trait Field5Functions {
   def fifth[S, A](implicit ev: Field5[S, A]): Lens[S, A] = ev.fifth
+}
+
+object Field5 extends Field5Functions{
+  /** lift an instance of [[Field5]] using an [[Iso]] */
+  def fromIso[S, A, B](iso: Iso[S, A])(implicit ev: Field5[A, B]): Field5[S, B] = new Field5[S, B] {
+    def fifth: Lens[S, B] =
+      iso composeLens ev.fifth
+  }
+
+  /************************************************************************************************/
+  /** Std instances                                                                               */
+  /************************************************************************************************/
+
+  implicit def tuple5Field5[A1, A2, A3, A4, A5]: Field5[(A1, A2, A3, A4, A5), A5] = new Field5[(A1, A2, A3, A4, A5), A5] {
+    def fifth = Lens((_: (A1, A2, A3, A4, A5))._5)(a => t => t.copy(_5 = a))
+  }
+
+  implicit def tuple6Field5[A1, A2, A3, A4, A5, A6]: Field5[(A1, A2, A3, A4, A5, A6), A5] = new Field5[(A1, A2, A3, A4, A5, A6), A5] {
+    def fifth = Lens((_: (A1, A2, A3, A4, A5, A6))._5)(a => t => t.copy(_5 = a))
+  }
 }
