+++ core/shared/src/main/scala/monocle/function/FilterIndex.scala
@@ -18,19 +18,8 @@ abstract class FilterIndex[S, I, A] extends Serializable {
   def filterIndex(predicate: I => Boolean): Traversal[S, A]
 }
 
-object FilterIndex extends FilterIndexFunctions {
-  /** lift an instance of [[FilterIndex]] using an [[Iso]] */
-  def fromIso[S, A, I, B](iso: Iso[S, A])(implicit ev: FilterIndex[A, I, B]): FilterIndex[S, I, B] = new FilterIndex[S, I, B] {
-    override def filterIndex(predicate: I => Boolean): Traversal[S, B] =
-      iso composeTraversal ev.filterIndex(predicate)
-  }
-}
-
 trait FilterIndexFunctions {
-
-  def filterIndex[S, I, A](predicate: I => Boolean)
-                            (implicit ev: FilterIndex[S, I, A]): Traversal[S, A] = ev.filterIndex(predicate)
-
+  def filterIndex[S, I, A](predicate: I => Boolean)(implicit ev: FilterIndex[S, I, A]): Traversal[S, A] = ev.filterIndex(predicate)
 
   def traverseFilterIndex[S[_]: Traverse, A](zipWithIndex: S[A] => S[(A, Int)]): FilterIndex[S[A], Int, A] = new FilterIndex[S[A], Int, A]{
     def filterIndex(predicate: Int => Boolean) = new Traversal[S[A], A] {
@@ -38,5 +27,64 @@ trait FilterIndexFunctions {
         zipWithIndex(s).traverse { case (a, j) => if(predicate(j)) f(a) else Applicative[F].point(a) }
     }
   }
+}
 
+object FilterIndex extends FilterIndexFunctions {
+  /** lift an instance of [[FilterIndex]] using an [[Iso]] */
+  def fromIso[S, A, I, B](iso: Iso[S, A])(implicit ev: FilterIndex[A, I, B]): FilterIndex[S, I, B] = new FilterIndex[S, I, B] {
+    def filterIndex(predicate: I => Boolean): Traversal[S, B] =
+      iso composeTraversal ev.filterIndex(predicate)
+  }
+
+  /************************************************************************************************/
+  /** Std instances                                                                               */
+  /************************************************************************************************/
+  import scalaz.std.list._
+  import scalaz.std.stream._
+  import scalaz.std.vector._
+
+  implicit def listFilterIndex[A]: FilterIndex[List[A], Int, A] =
+    traverseFilterIndex(_.zipWithIndex)
+
+  implicit def mapFilterIndex[K, V]: FilterIndex[Map[K,V], K, V] = new FilterIndex[Map[K, V], K, V] {
+    import scalaz.syntax.applicative._
+    def filterIndex(predicate: K => Boolean) = new Traversal[Map[K, V], V] {
+      def modifyF[F[_]: Applicative](f: V => F[V])(s: Map[K, V]): F[Map[K, V]] =
+        s.toList.traverse{ case (k, v) =>
+          (if(predicate(k)) f(v) else v.point[F]).strengthL(k)
+        }.map(_.toMap)
+    }
+  }
+
+  implicit def streamFilterIndex[A]: FilterIndex[Stream[A], Int, A] =
+    traverseFilterIndex(_.zipWithIndex)
+
+  implicit val stringFilterIndex: FilterIndex[String, Int, Char] = new FilterIndex[String, Int, Char]{
+    def filterIndex(predicate: Int => Boolean) =
+      monocle.std.string.stringToList composeTraversal FilterIndex.filterIndex[List[Char], Int, Char](predicate)
+  }
+
+  implicit def vectorFilterIndex[A]: FilterIndex[Vector[A], Int, A] =
+    traverseFilterIndex(_.zipWithIndex)
+
+  /************************************************************************************************/
+  /** Scalaz instances                                                                            */
+  /************************************************************************************************/
+  import scalaz.{==>>, IList, NonEmptyList, Order}
+
+  implicit def iListFilterIndex[A]: FilterIndex[IList[A], Int, A] =
+    traverseFilterIndex(_.zipWithIndex)
+
+  implicit def iMapFilterIndex[K: Order, V]: FilterIndex[K ==>> V, K, V] = new FilterIndex[K ==>> V, K, V] {
+    import scalaz.syntax.applicative._
+    def filterIndex(predicate: K => Boolean) = new Traversal[K ==>> V, V] {
+      def modifyF[F[_]: Applicative](f: V => F[V])(s: K ==>> V): F[K ==>> V] =
+        s.toList.traverse{ case (k, v) =>
+          (if(predicate(k)) f(v) else v.point[F]).strengthL(k)
+        }.map(==>>.fromList(_))
+    }
+  }
+
+  implicit def nelFilterIndex[A]: FilterIndex[NonEmptyList[A], Int, A] =
+    traverseFilterIndex(_.zipWithIndex)
 }
