+++ core/shared/src/main/scala/monocle/function/Plated.scala
@@ -3,10 +3,10 @@ package monocle.function
 import monocle.{Setter, Traversal}
 
 import scala.annotation.implicitNotFound
-import scalaz.{Monad, State}
 import scalaz.std.stream._
-import scalaz.syntax.monad._
 import scalaz.std.anyVal._
+import scalaz.syntax.monad._
+import scalaz.{Applicative, Monad, State, Traverse}
 
 /**
   * [[Plated]] is a type-class for types which can extract their immediate
@@ -19,8 +19,6 @@ import scalaz.std.anyVal._
 abstract class Plated[A] extends Serializable { self =>
   def plate: Traversal[A, A]
 }
-
-object Plated extends PlatedFunctions
 
 trait PlatedFunctions {
 
@@ -42,7 +40,7 @@ trait PlatedFunctions {
     * a fixpoint (this is an infinite loop if there is no fixpoint)
     */
   def rewrite[A: Plated](f: A => Option[A])(a: A): A =
-    rewriteOf(plate[A].asSetter)(f)(a)
+  rewriteOf(plate[A].asSetter)(f)(a)
 
   /**
     * rewrite a target by applying a rule within a [[Setter]], as often as
@@ -59,11 +57,11 @@ trait PlatedFunctions {
 
   /** transform every element */
   def transform[A: Plated](f: A => A)(a: A): A =
-    transformOf(plate[A].asSetter)(f)(a)
+  transformOf(plate[A].asSetter)(f)(a)
 
   /** transform every element by applying a [[Setter]] */
   def transformOf[A](l: Setter[A, A])(f: A => A)(a: A): A =
-    l.modify(b => transformOf(l)(f)(f(b)))(a)
+  l.modify(b => transformOf(l)(f)(f(b)))(a)
 
   /** transforming counting changes */
   def transformCounting[A: Plated](f: A => Option[A])(a: A): (Int, A) = {
@@ -82,4 +80,89 @@ trait PlatedFunctions {
     go(a)
   }
 
+}
+
+object Plated extends PlatedFunctions {
+  /************************************************************************************************/
+  /** Std instances                                                                               */
+  /************************************************************************************************/
+
+  implicit def listPlated[A]: Plated[List[A]] = new Plated[List[A]] {
+    val plate: Traversal[List[A], List[A]] = new Traversal[List[A], List[A]] {
+      def modifyF[F[_]: Applicative](f: List[A] => F[List[A]])(s: List[A]): F[List[A]] =
+        s match {
+          case x :: xs => Applicative[F].map(f(xs))(x :: _)
+          case Nil => Applicative[F].point(Nil)
+        }
+    }
+  }
+
+  implicit def streamPlated[A]: Plated[Stream[A]] = new Plated[Stream[A]] {
+    val plate: Traversal[Stream[A], Stream[A]] = new Traversal[Stream[A], Stream[A]] {
+      def modifyF[F[_]: Applicative](f: Stream[A] => F[Stream[A]])(s: Stream[A]): F[Stream[A]] =
+        s match {
+          case x #:: xs => Applicative[F].map(f(xs))(x #:: _)
+          case Stream() => Applicative[F].point(Stream.empty)
+        }
+    }
+  }
+
+  implicit val stringPlated: Plated[String] = new Plated[String] {
+    val plate: Traversal[String, String] = new Traversal[String, String] {
+      def modifyF[F[_]: Applicative](f: String => F[String])(s: String): F[String] =
+        s.headOption match {
+          case Some(h) => Applicative[F].map(f(s.tail))(h.toString ++ _)
+          case None => Applicative[F].point("")
+        }
+    }
+  }
+
+  implicit def vectorPlated[A]: Plated[Vector[A]] = new Plated[Vector[A]] {
+    val plate: Traversal[Vector[A], Vector[A]] = new Traversal[Vector[A], Vector[A]] {
+      def modifyF[F[_]: Applicative](f: Vector[A] => F[Vector[A]])(s: Vector[A]): F[Vector[A]] =
+        s match {
+          case h +: t => Applicative[F].map(f(t))(h +: _)
+          case _ => Applicative[F].point(Vector.empty)
+        }
+    }
+  }
+
+  /************************************************************************************************/
+  /** Scalaz instances                                                                            */
+  /************************************************************************************************/
+  import scalaz.{Cofree, Free, IList, ICons, INil, Tree}
+
+  implicit def cofreePlated[S[_]: Traverse, A]: Plated[Cofree[S, A]] = new Plated[Cofree[S, A]] {
+    val plate: Traversal[Cofree[S, A], Cofree[S, A]] = new Traversal[Cofree[S, A], Cofree[S, A]] {
+      def modifyF[F[_]: Applicative](f: Cofree[S, A] => F[Cofree[S, A]])(s: Cofree[S, A]): F[Cofree[S, A]] =
+        Applicative[F].map(Traverse[S].traverse(s.t.run)(f))(Cofree(s.head, _))
+    }
+  }
+
+  implicit def freePlated[S[_]: Traverse, A]: Plated[Free[S, A]] = new Plated[Free[S, A]] {
+    val plate: Traversal[Free[S, A], Free[S, A]] = new Traversal[Free[S, A], Free[S, A]] {
+      def modifyF[F[_]: Applicative](f: Free[S, A] => F[Free[S, A]])(s: Free[S, A]): F[Free[S, A]] =
+        s.resume.fold(
+          as => Applicative[F].map(Traverse[S].traverse(as)(f))(Free.roll),
+          x => Applicative[F].point(Free.point(x))
+        )
+    }
+  }
+
+  implicit def ilistPlated[A]: Plated[IList[A]] = new Plated[IList[A]] {
+    val plate: Traversal[IList[A], IList[A]] = new Traversal[IList[A], IList[A]] {
+      def modifyF[F[_]: Applicative](f: IList[A] => F[IList[A]])(s: IList[A]): F[IList[A]] =
+        s match {
+          case ICons(x, xs) => Applicative[F].map(f(xs))(x :: _)
+          case INil() => Applicative[F].point(INil())
+        }
+    }
+  }
+
+  implicit def treePlated[A]: Plated[Tree[A]] = new Plated[Tree[A]] {
+    val plate: Traversal[Tree[A], Tree[A]] = new Traversal[Tree[A], Tree[A]] {
+      def modifyF[F[_]: Applicative](f: Tree[A] => F[Tree[A]])(s: Tree[A]): F[Tree[A]] =
+        Applicative[F].map(Traverse[Stream].traverse(s.subForest)(f))(Tree.Node(s.rootLabel, _))
+    }
+  }
 }
