+++ core/shared/src/main/scala/monocle/function/Snoc.scala
@@ -1,10 +1,10 @@
 package monocle.function
 
 import monocle.function.fields._
-import monocle.std.tuple2._
 import monocle.{Iso, Optional, Prism}
 
 import scala.annotation.implicitNotFound
+import scalaz.{Applicative, \/}
 
 /**
  * Typeclass that defines a [[Prism]] between an `S` and its init `S` and last `S`
@@ -20,14 +20,6 @@ abstract class Snoc[S, A] extends Serializable {
   def lastOption: Optional[S, A] = snoc composeLens second
 }
 
-object Snoc extends SnocFunctions {
-  /** lift an instance of [[Snoc]] using an [[Iso]] */
-  def fromIso[S, A, B](iso: Iso[S, A])(implicit ev: Snoc[A, B]): Snoc[S, B] = new Snoc[S, B] {
-    override def snoc: Prism[S, (S, B)] =
-      iso composePrism ev.snoc composeIso iso.reverse.first
-  }
-}
-
 trait SnocFunctions {
   final def snoc[S, A](implicit ev: Snoc[S, A]): Prism[S, (S, A)] = ev.snoc
 
@@ -36,9 +28,66 @@ trait SnocFunctions {
 
   /** append an element to the end */
   final def _snoc[S, A](init: S, last: A)(implicit ev: Snoc[S, A]): S =
-    ev.snoc.reverseGet((init, last))
+  ev.snoc.reverseGet((init, last))
 
   /** deconstruct an S between its init and last */
   final def _unsnoc[S, A](s: S)(implicit ev: Snoc[S, A]): Option[(S, A)] =
-    ev.snoc.getOption(s)
+  ev.snoc.getOption(s)
+}
+
+object Snoc extends SnocFunctions {
+  /** lift an instance of [[Snoc]] using an [[Iso]] */
+  def fromIso[S, A, B](iso: Iso[S, A])(implicit ev: Snoc[A, B]): Snoc[S, B] = new Snoc[S, B] {
+    def snoc: Prism[S, (S, B)] =
+      iso composePrism ev.snoc composeIso iso.reverse.first
+  }
+
+  /************************************************************************************************/
+  /** Std instances                                                                               */
+  /************************************************************************************************/
+  import scalaz.std.option._
+
+  implicit def listSnoc[A]: Snoc[List[A], A] = new Snoc[List[A], A]{
+    def snoc = Prism[List[A], (List[A], A)](
+      s => Applicative[Option].apply2(\/.fromTryCatchNonFatal(s.init).toOption, s.lastOption)((_,_))){
+      case (init, last) => init :+ last
+    }
+  }
+
+  implicit def streamSnoc[A]: Snoc[Stream[A], A] = new Snoc[Stream[A], A]{
+    def snoc = Prism[Stream[A], (Stream[A], A)]( s =>
+      for {
+        init <- if(s.isEmpty) None else Some(s.init)
+        last <- s.lastOption
+      } yield (init, last)){
+      case (init, last) => init :+ last
+    }
+  }
+
+  implicit val stringSnoc: Snoc[String, Char] = new Snoc[String, Char]{
+    def snoc =
+      Prism[String, (String, Char)](
+        s => if(s.isEmpty) None else Some((s.init, s.last))){
+        case (init, last) => init :+ last
+      }
+  }
+
+  implicit def vectorSnoc[A]: Snoc[Vector[A], A] = new Snoc[Vector[A], A]{
+    def snoc = Prism[Vector[A], (Vector[A], A)](
+      v => if(v.isEmpty) None else Some((v.init, v.last))){
+      case (xs, x) => xs :+ x
+    }
+  }
+
+  /************************************************************************************************/
+  /** Scalaz instances                                                                            */
+  /************************************************************************************************/
+  import scalaz.IList
+
+  implicit def iListSnoc[A]: Snoc[IList[A], A] = new Snoc[IList[A], A]{
+    def snoc = Prism[IList[A], (IList[A], A)](
+      il => Applicative[Option].apply2(il.initOption, il.lastOption)((_,_))){
+      case (init, last) => init :+ last
+    }
+  }
 }
