+++ core/shared/src/main/scala/monocle/std/Set.scala
@@ -1,18 +1,7 @@
 package monocle.std
 
-import monocle.function.{At, Empty}
-import monocle.{Lens, Prism}
+@deprecated("instances have been move to typeclass companion object", since = "1.4.0")
+object set
 
-object set extends SetOptics
-
-trait SetOptics {
-
-  implicit def emptySet[A]: Empty[Set[A]] = new Empty[Set[A]] {
-    def empty = Prism[Set[A], Unit](s => if(s.isEmpty) Some(()) else None)(_ => Set.empty[A])
-  }
-
-  implicit def atSet[A]: At[Set[A], A, Boolean] = new At[Set[A], A, Boolean] {
-    def at(a: A) = Lens[Set[A], Boolean](_.contains(a))(b => set => if(b) set + a else set - a)
-  }
-
-}
+@deprecated("instances have been move to typeclass companion object", since = "1.4.0")
+trait SetOptics
