+++ core/shared/src/main/scala/monocle/std/String.scala
@@ -1,10 +1,7 @@
 package monocle.std
 
-import monocle.function._
-import monocle.std.list._
-import monocle.{Iso, Prism, Traversal}
+import monocle.{Iso, Prism}
 
-import scalaz.Applicative
 import scalaz.std.list._
 import scalaz.std.option._
 import scalaz.syntax.traverse._
@@ -27,59 +24,6 @@ trait StringOptics {
 
   val stringToByte: Prism[String, Byte] =
     stringToLong composePrism long.longToByte
-
-
-  implicit val stringEmpty: Empty[String] =
-    new Empty[String] {
-      def empty = Prism[String, Unit](s => if(s.isEmpty) Some(()) else None)(_ => "")
-    }
-
-  implicit val stringReverse: Reverse[String, String] =
-    Reverse.reverseFromReverseFunction[String](_.reverse)
-
-  implicit val stringEach: Each[String, Char] =
-    new Each[String, Char] {
-      def each =
-        stringToList composeTraversal Each.each[List[Char], Char]
-    }
-
-  implicit val stringIndex: Index[String, Int, Char] =
-    new Index[String, Int, Char]{
-      def index(i: Int) =
-        stringToList composeOptional Index.index[List[Char], Int, Char](i)
-    }
-
-  implicit val stringFilterIndex: FilterIndex[String, Int, Char] =
-    new FilterIndex[String, Int, Char]{
-      def filterIndex(predicate: Int => Boolean) =
-        stringToList composeTraversal FilterIndex.filterIndex[List[Char], Int, Char](predicate)
-    }
-
-  implicit val stringCons: Cons[String, Char] =
-    new Cons[String, Char] {
-      def cons =
-        Prism[String, (Char, String)](s =>
-          if(s.isEmpty) None else Some((s.head, s.tail))
-        ){ case (h, t) => h + t }
-  }
-
-  implicit val stringSnoc: Snoc[String, Char] = new Snoc[String, Char]{
-    def snoc =
-      Prism[String, (String, Char)](
-        s => if(s.isEmpty) None else Some((s.init, s.last))){
-        case (init, last) => init :+ last
-      }
-  }
-
-  implicit val stringPlated: Plated[String] = new Plated[String] {
-    val plate: Traversal[String, String] = new Traversal[String, String] {
-      def modifyF[F[_]: Applicative](f: String => F[String])(s: String): F[String] =
-        s.headOption match {
-          case Some(h) => Applicative[F].map(f(s.tail))(h.toString ++ _)
-          case None => Applicative[F].point("")
-        }
-    }
-  }
 
   private def parseLong(s: String): Option[Long] = {
     // we reject cases where String will be an invalid Prism according 2nd Prism law
