+++ core/shared/src/main/scala/monocle/std/Tree.scala
@@ -1,15 +1,12 @@
 package monocle.std
 
-import monocle.function._
-import monocle.{Iso, Lens, Traversal}
+import monocle.Lens
 import monocle.function.all._
-import monocle.std.stream._
 
-import scalaz.Tree.{Leaf, Node}
-import scalaz.{Applicative, Traverse, Tree}
-import scalaz.std.stream._
 import scala.annotation.tailrec
 import scala.collection.immutable.Stream.Empty
+import scalaz.Tree
+import scalaz.Tree.{Leaf, Node}
 
 object tree extends TreeOptics
 
@@ -49,20 +46,6 @@ trait TreeOptics {
     }
 
     Lens(_get)(_set)
-  }
-
-  implicit def treeEach[A]: Each[Tree[A], A] = Each.traverseEach[Tree, A]
-
-  implicit def treeReverse[A]: Reverse[Tree[A], Tree[A]] = new Reverse[Tree[A], Tree[A]] {
-    def reverse = Iso[Tree[A], Tree[A]](reverseTree)(reverseTree)
-    private def reverseTree(tree: Tree[A]): Tree[A] = Node(tree.rootLabel, tree.subForest.reverse.map(reverseTree))
-  }
-
-  implicit def treePlated[A]: Plated[Tree[A]] = new Plated[Tree[A]] {
-    val plate: Traversal[Tree[A], Tree[A]] = new Traversal[Tree[A], Tree[A]] {
-      def modifyF[F[_]: Applicative](f: Tree[A] => F[Tree[A]])(s: Tree[A]): F[Tree[A]] =
-        Applicative[F].map(Traverse[Stream].traverse(s.subForest)(f))(Node(s.rootLabel, _))
-    }
   }
 
 }
