+++ test/shared/src/test/scala/monocle/function/Cons1Spec.scala
@@ -3,12 +3,12 @@ package monocle.function
 import monocle.MonocleSuite
 import monocle.law.discipline.function.Cons1Tests
 
-import scalaz.NonEmptyList
+import scalaz.IList
 
 class Cons1Spec extends MonocleSuite {
 
-  implicit val clistCons1: Cons1[CNel, Char, List[Char]] = Cons1.fromIso(CNel.toNel)
+  implicit val clistCons1: Cons1[CNel, Char, IList[Char]] = Cons1.fromIso(CNel.toNel)
 
-  checkAll("fromIso", Cons1Tests[CNel, Char, List[Char]])
+  checkAll("fromIso", Cons1Tests[CNel, Char, IList[Char]])
 
 }
