+++ test/shared/src/test/scala/monocle/std/NonEmptyListSpec.scala
@@ -4,7 +4,7 @@ import monocle.MonocleSuite
 import monocle.law.discipline.IsoTests
 import monocle.law.discipline.function._
 
-import scalaz.NonEmptyList
+import scalaz.{IList, NonEmptyList}
 
 class NonEmptyListSpec extends MonocleSuite {
   checkAll("nelToAndOne", IsoTests(nelToOneAnd[Int]))
@@ -14,6 +14,6 @@ class NonEmptyListSpec extends MonocleSuite {
   checkAll("index NonEmptyList", IndexTests[NonEmptyList[Int], Int, Int])
   checkAll("filterIndex NonEmptyList", FilterIndexTests[NonEmptyList[Int], Int, Int])
   checkAll("reverse NonEmptyList", ReverseTests[NonEmptyList[Int]])
-  checkAll("cons1 NonEmptyList", Cons1Tests[NonEmptyList[Int], Int, List[Int]])
-  checkAll("snoc1 NonEmptyList", Snoc1Tests[NonEmptyList[Int], List[Int], Int])
+  checkAll("cons1 NonEmptyList", Cons1Tests[NonEmptyList[Int], Int, IList[Int]])
+  checkAll("snoc1 NonEmptyList", Snoc1Tests[NonEmptyList[Int], IList[Int], Int])
 }
